third_party/WebKit/LayoutTests/virtual/feature-policy/http/tests/feature-policy/fullscreen-allowed-by-container-policy-expected.txt

This is a testharness.js-based test.
PASS Fullscreen is allowed on URL: resources/feature-policy-fullscreen.html with allowfullscreen = false
PASS Fullscreen is allowed on URL: http://localhost:8000/feature-policy/resources/feature-policy-fullscreen.html with allowfullscreen = false
PASS Fullscreen is allowed on URL: resources/feature-policy-fullscreen.html with allowfullscreen = true
PASS Fullscreen is allowed on URL: http://localhost:8000/feature-policy/resources/feature-policy-fullscreen.html with allowfullscreen = true
Harness: the test ran to completion.
